update to support axios v0.18.0
The most important change in this commit is the digest function. In
version 0.13.0 there was a
[commit in axios](https://github.com/axios/axios/commit/e833a2f7e4a5ec32516e3f17763374c0377a4323)
moving the request transformers to after the request interceptors. This
means that the request interceptor will see a different confguration
from the response interceptor, so the output of the digest function will
be different in them, breaking axios-vcr functionality.
To solve this we can simply apply the same request transformers to the
configuration in the request interceptor.